Refactored init db to database package.
