+++ database/database.go
@@ -1,6 +1,7 @@
-package main
+package database
 
 import (
+	"fmt"
 	"os"
 
 	_ "github.com/go-sql-driver/mysql"
@@ -18,13 +19,14 @@ const (
 		PRIMARY KEY (id)
 	) ENGINE=InnoDB AUTO_INCREMENT=1 DEFAULT CHARSET=utf8;`
 	checkIfSchemaExists = `SELECT COUNT(*) as does_exist FROM INFORMATION_SCHEMA.TABLES WHERE TABLE_NAME = 'visits'`
+	insertVisit         = `INSERT INTO visits (ip, host, path, user_agent, created_at) VALUES (:ip, :host, :path, :user_agent, :created_at)`
 )
 
 type TableCheck struct {
 	DoesExist int `db:"does_exist"`
 }
 
-func VerifySchema() {
+func InitializeDatabase() *sqlx.DB {
 	db := sqlx.MustConnect("mysql", os.Getenv("PING_DB"))
 	db.Ping()
 	var check TableCheck
@@ -35,8 +37,22 @@ func VerifySchema() {
 	if check.DoesExist < 1 {
 		db.MustExec(schema)
 	}
+	return db
 }
 
-func main() {
-	VerifySchema()
+type Visit struct {
+	IP        string `db:"ip"`
+	Host      string `db:"host"`
+	Path      string `db:"path"`
+	UserAgent string `db:"user_agent"`
+	CreatedAt string `db:"created_at"`
+}
+
+func (v *Visit) String() string {
+	return fmt.Sprintf("<%s | %s requested %s%s @ %s>", v.CreatedAt, v.IP, v.Host, v.Path, v.UserAgent)
+}
+
+func (v *Visit) Save(db *sqlx.DB) error {
+	_, err := db.NamedExec(insertVisit, v)
+	return err
 }
